% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bal.long.otus}
\alias{bal.long.otus}
\title{Simulated OTU data for balanced longitudinal study design.}
\format{A matrix with 15 rows and 10 columns. Rows are samples, columns are OTUs.}
\usage{
data(bal.long.otus)
}
\description{
Simulation code is included in the package vignette. 
Corresponding metadata is stored in `bal.long.meta`.
}
\keyword{datasets}
