% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(otus, metadata, paired)
}
\arguments{
\item{otus}{Matrix of OTU counts or proportions. Notes: (1) Will be transformed to 
proportions if it's not already; (2) Row names must be sample identifiers 
(matching metadata), and column names must be OTU identifiers (enforced if 
using UniFrac distances).}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (if using longitudinal distances, this must be numeric or 
convertable to numeric).}

\item{paired}{Logical indicating whether to use the paired version of the metric (TRUE) or the 
longitudinal version (FALSE). Paired analyis is only possible when there are exactly 2 
unique time points/identifiers for each subject or pair.}
}
\value{
Returns checked (and possibly fixed) OTU and metadata files.
}
\description{
Checks data input for transformation (pltransform) and dissimilarity (pldist) functions.
}
