% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltransform.R
\name{counts2props}
\alias{counts2props}
\title{counts2props}
\usage{
counts2props(x)
}
\arguments{
\item{x}{Matrix of OTU counts (rows are subjects, columns are taxa).}
}
\value{
n x p matrix of OTU proportions.
}
\description{
Converts OTU counts to OTU proportions/relative abundances.
}
