% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{data_prep}
\usage{
data_prep(otus, metadata, paired, pseudoct = NULL)
}
\arguments{
\item{otus}{Matrix of OTU counts (better) or proportions. Row names must be sample identifiers 
(matching metadata), and column names must be OTU identifiers (enforced if 
using UniFrac distances).}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (if using longitudinal distances, this must be numeric or 
convertable to numeric).}

\item{paired}{Logical indicating whether to use the paired version of the metric (TRUE) or the 
longitudinal version (FALSE). Paired analyis is only possible when there are exactly 2 
unique time points/identifiers for each subject or pair.}

\item{pseudoct}{Pseudocount value to be added to each cell of the matrix. Default is NULL; if NULL, 
0.5 will be added if data are counts, min(1e-06, 0.5*min(nonzero p)) will be added if data are 
proportions, and nothing will be added if no cells have zero values. This is only done for the 
CLR-transformed data matrix, not the OTU proportion matrix.}
}
\value{
Returns OTU proportions, CLR-transformed OTU proportions, and metadata files, all checked 
    for formatting and value problems.
}
\description{
Checks data input for transformation (pltransform) and dissimilarity (pldist) functions. Prepares 
matrix of OTU proportions and CLR-transformed matrix.
}
