% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_jaccard.R
\name{flexsign}
\alias{flexsign}
\title{flexsign}
\usage{
flexsign(v1, v2)
}
\arguments{
\item{v1}{First vector}

\item{v2}{Second vector}
}
\value{
Returns an n x n distance matrix.
}
\description{
Sign function that considers 0 both positive and negative. Returns 1 if the two numbers are the same sign, 0 otherwise. Vectorized (compares vectors elementwise).
}
