% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_kulczynski.R
\name{kulczynski}
\alias{kulczynski}
\title{Paired or longitudinal Kulczynski distances}
\usage{
kulczynski(tsf.data, paired, binary)
}
\arguments{
\item{tsf.data}{Transformed OTU table and metadata (from function pl.transform)}

\item{paired}{Logical indicating whether paired analysis is desired}

\item{binary}{Logical indicating whether to use the binary version of the distance}
}
\value{
Returns an n x n distance matrix.
}
\description{
The distances are calculated as follows, where d_k^X is the within-subject 
    measure of change appropriate to the setting (paired/longitudinal and 
    quantitative/qualitative), as described in the full package documentation 
    and vignette. 
    
Paired, qualitative: \eqn{D_{AB} = 1 - (1/m) \sum_k I[d_k^A = d_k^B] I[d_k^A \neq 0]} 
Paired, quantitative: \eqn{D_{AB} = 1 - (2/m) \sum_k \min(|d_k^A|, |d_k^B|) I[sgn(d_k^A) = sgn(d_k^B)]}
Longitudinal: \eqn{D_{AB} = 1 - (1/m) * \sum_k \min(d_k^A, d_k^B)}
}
