% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paired.meta}
\alias{paired.meta}
\title{Simulated metadata for paired study design.}
\format{A data frame with 10 rows and 3 columns. 
\describe{
  \item{subjID}{Subject identifiers}
  \item{sampID}{Sample identifiers, matches row names of OTU count matrix}
  \item{time}{Time indicator, takes values 1 or 2}
}}
\usage{
data(paired.meta)
}
\description{
Simulation code is included in the package vignette. 
Corresponding OTU matrix is stored in `paired.otus`.
}
\keyword{datasets}
