% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paired.otus}
\alias{paired.otus}
\title{Simulated OTU data for paired study design.}
\format{A matrix with 10 rows and 10 columns. Rows are samples, columns are OTUs.}
\usage{
data(paired.otus)
}
\description{
Simulation code is included in the package vignette. 
Corresponding metadata is stored in `paired.meta`.
}
\keyword{datasets}
