% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{psct_clr}
\alias{psct_clr}
\title{psct_clr}
\usage{
psct_clr(otus, pseudocount = NULL)
}
\arguments{
\item{otus}{Matrix of OTU counts or proportions. Rows are samples, columns are OTUs.}

\item{pseudocount}{Pseudocount to be added to all values in OTU matrix. If NULL, then 
0.5 is added if entries are counts, or min(1e-06, 0.5*min(nonzero p)) if entries
are proportions. If all entries are nonzero, nothing is added.}
}
\value{
Matrix of CLR-transformed data (dimensions, rownames, colnames match input)
}
\description{
Adds a pseudocount, (re-)closes to proportions, and then applies centered log-ratio transformation
}
