% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_unifrac.R
\name{PUniFrac}
\alias{PUniFrac}
\title{PUniFrac}
\usage{
PUniFrac(otu.tab, tree, gam = c(0, 0.5, 1), metadata)
}
\arguments{
\item{otu.tab}{OTU count table, containing 2*n rows (samples) and q columns (OTUs)}

\item{tree}{Rooted phylogenetic tree of R class "phylo"}

\item{gam}{Parameter controlling weight on abundant lineages. The same weight is used within a subject as between subjects.}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (variable with two unique levels, column name "time").}
}
\value{
Returns a (K+1) dimensional array containing the longitudinal UniFrac dissimilarities 
    with the K specified gamma values plus the unweighted distance. The unweighted dissimilarity 
    matrix may be accessed by result[,,"d_UW"], and the generalized dissimilarities by result[,,"d_G"] 
    where G is the particular choice of gamma.
}
\description{
Paired UniFrac distances for comparing changes in
microbial communities across 2 groups or time points.
}
\details{
Based in part on Jun Chen & Hongzhe Li (2012), GUniFrac.

Computes difference between time points and then calculates
difference of these differences, resulting in a dissimilarity
matrix that can be used in a variety of downstream 
distance-based analyses.
}
\examples{
## Example: simulated tree and OTU table
ntaxa = 5; nsubj = 10
library(ape)
sim.tree = rtree(n=ntaxa)  # simulated rooted phylogenetic tree
sim.otu <- matrix(runif(2 x ntaxa x nsubj, 0, 100), nrow = 2 x nsubj)
for (i in 1:nrow(sim.otu)) { sim.otu[i, ] <- sim.otu[i, ] / sum(sim.otu[i, ]) }
rownames(sim.otu) <- paste("Subj", rep(1:nsubj, each = 2), "_Time", rep(1:2, nsubj), sep = "")  # Row names must match samples in metadata
colnames(sim.otu) <- paste("t", 1:ntaxa, sep = "")   # Column names must match tree tip labels
sim.meta <- data.frame(subjID = rep(paste("Subj", 1:nsubj, sep = ""), each = 2), sampID = rownames(sim.otu), time = rep(c(1,2), nsubj))
Ds <- PUniFrac(otu.tab = sim.otu, tree = sim.tree, gam = c(0, 0.5, 1), metadata = sim.meta)
D_unweighted <- Ds[,,"d_UW"]   ## access individual array elements
D_gen_a0.5 <- Ds[,,"d_0.5"]

}
