% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.tree}
\alias{sim.tree}
\title{Simulated rooted phylogenetic tree.}
\format{An object of class "phylo".}
\usage{
data(sim.tree)
}
\description{
Simulation code is included in the package vignette. 
Tree includes 10 OTUs and may be used with any of the 
simulated data sets (paired, balanced longitudinal, or 
unbalanced longitudinal).
}
\keyword{datasets}
