% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltransform.R
\name{tsf_paired}
\alias{tsf_paired}
\title{tsf_paired}
\usage{
tsf_paired(otu.props, otu.clr, metadata, norm = TRUE)
}
\arguments{
\item{otu.props}{Matrix of OTU proportions (rownames = sample IDs, colnames = OTU IDs)}

\item{otu.clr}{Matrix of CLR-transformed OTU proportions (rownames = sample IDs, colnames = OTU IDs)}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (must have two unique values for paired transformation).}

\item{norm}{Indicator of whether to normalize the difference to average taxon abundance or not (default TRUE)}
}
\value{
List with the following elements. Data matrices have subject identifiers 
    as row names and OTU identifiers as column names.  
    \item{dat.binary}{n x p matrix of data after paired, binary/qualitative transformation} 
    \item{dat.quant.prop}{n x p matrix of data after paired, quantitative transformation applied to OTU proportions} 
    \item{dat.quant.clr}{n x p matrix of data after paired, quantitative transformation applied to CLR-transformed OTU proportions} 
    \item{avg.prop}{n x p matrix with overall average proportion of each taxon}
}
\description{
OTU transformation for paired data. Computes within-subject change (in presence 
    for qualitative metrics and abundance for quantitative metrics) between time 
    points for each taxon.
}
