% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unbal.long.meta}
\alias{unbal.long.meta}
\title{Simulated metadata for balanced longitudinal study design.}
\format{A data frame with 14 rows and 3 columns. 
\describe{
  \item{subjID}{Subject identifiers}
  \item{sampID}{Sample identifiers, matches row names of OTU count matrix}
  \item{time}{Time indicator}
}}
\usage{
data(unbal.long.meta)
}
\description{
Simulation code is included in the package vignette. 
Corresponding OTU matrix is stored in `unbal.long.otus`.
}
\keyword{datasets}
