% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unbal.long.otus}
\alias{unbal.long.otus}
\title{Simulated OTU data for unbalanced longitudinal study design.}
\format{A matrix with 14 rows and 10 columns. Rows are samples, columns are OTUs.}
\usage{
data(unbal.long.otus)
}
\description{
Simulation code is included in the package vignette. 
Corresponding metadata is stored in `unbal.long.meta`.
}
\keyword{datasets}
