% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltransform.R
\name{tsf_long}
\alias{tsf_long}
\title{tsf_paired}
\usage{
tsf_long(otus, metadata)
}
\arguments{
\item{otus}{Matrix of OTU counts or proportions. Notes: (1) Will be transformed to 
proportions if it's not already; (2) Row names must be sample identifiers 
(matching metadata), and column names must be OTU identifiers (enforced if 
using UniFrac distances).}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (if using longitudinal distances, this must be numeric or 
convertable to numeric).}
}
\value{
List with the following elements. Both data matrices have subject identifiers 
    as row names and OTU identifiers as column names.  
    \item{dat.binary}{n x p matrix of data after longitudinal, binary/qualitative transformation} 
    \item{dat.quant}{n x p matrix of data after longitudinal, quantitative transformation} 
    \item{avg.prop}{n x p matrix with overall average proportion of each taxon}
}
\description{
OTU transformation for longitudinal data. Computes average within-subject change 
    (in presence for qualitative metrics, abundance for quantitative metrics) 
    during one unit of time for each taxon.
}
