% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltransform.R
\name{tsf_paired}
\alias{tsf_paired}
\title{tsf_paired}
\usage{
tsf_paired(otus, metadata)
}
\arguments{
\item{otus}{Matrix of OTU counts or proportions. Notes: (1) Will be transformed to 
proportions if it's not already; (2) Row names must be sample identifiers 
(matching metadata), and column names must be OTU identifiers (enforced if 
using UniFrac distances).}

\item{metadata}{Data frame with three columns: subject identifiers (n unique values, column name "subjID"), 
sample identifiers (must match row names of otu.tab, column name "sampID"), 
and time point or group identifier (must have two unique values for paired transformation).}
}
\value{
List with the following elements. Both data matrices have subject identifiers 
    as row names and OTU identifiers as column names.  
    \item{dat.binary}{n x p matrix of data after paired, binary/qualitative transformation} 
    \item{dat.quant}{n x p matrix of data after paired, quantitative transformation} 
    \item{avg.prop}{n x p matrix with overall average proportion of each taxon}
}
\description{
OTU transformation for paired data. Computes within-subject change (in presence 
    for qualitative metrics and abundance for quantitative metrics) between time 
    points for each taxon.
}
